selectDial for step nav, sep e2 controls handler
